Renamed the project's packages and (some of) its classes.
CBBF-40
