+++ bluebutton-server-app/src/test/java/gov/hhs/cms/bluebutton/server/app/RandomServerPortProvider.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.demo;
+package gov.hhs.cms.bluebutton.server.app;
 
 import java.io.IOException;
 import java.net.ServerSocket;
